# [8.4.0] Optimize RemoteOutputChecker.addOutputToDownload